Packaged up into full module file set   - endpoints now accessible at http://yourmagento.com/index.php/unity3d/index/categories http://yourmagento.com/index.php/unity3d/index/product http://yourmagento.com/index.php/unity3d/index/order
