+++ app/code/community/Zaeran/Unity3d/etc/config.xml
@@ -1,0 +1,19 @@
+<?xml version="1.0"?>
+<config>
+    <modules>
+        <Zaeran_Unity3d>
+            <version>0.1.1</version>    <!-- Version number of your module -->
+        </Zaeran_Unity3d>
+    </modules>
+    <frontend>
+        <routers>
+            <Zaeran_Unity3d>
+                <use>standard</use>
+                <args>
+                    <module>Zaeran_Unity3d</module>
+                    <frontName>unity3d</frontName>
+                </args>
+            </Zaeran_Unity3d>
+        </routers>
+    </frontend>
+</config>
